feat!: rename package to reshift
Rename the import package functions/ -> reshift/ and the distribution
odmd-subid-cpd -> reshift. Updates all imports (app, benchmarks, example
notebooks), coverage source, README/mkdocs branding, and the lockfile.

BREAKING CHANGE: import path changes from 'functions.*' to 'reshift.*';
PyPI/distribution name changes from 'odmd-subid-cpd' to 'reshift'.
