+++ examples/__init__.py
@@ -1,1 +1,1 @@
-"""Example notebooks and configuration for the odmd-subid-cpd package."""
+"""Example notebooks and configuration for the reshift package."""
