+++ reshift/rolling.py
